pyproject.toml
[project]
name = "djangorest-starter-kit"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikita Lugachev",email = "nikitalugachev149th@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.7)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.11)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "poetry-core (>=2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
